requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-pinecone
langchain-community
langgraph
faiss-cpu
pydantic
python-dotenv
httpx
pytest
pinecone-client
mcp>=1.0.0
httpx>=0.25.0
PyPDF2
pdfplumber
python-multipart